move `DebugAsType` into its own module
