+++ crates/but/src/utils/mod.rs
@@ -10,6 +10,9 @@ pub use object_id::{shorten_hex_object_id, shorten_object_id, split_short_id};
 
 mod pager;
 
+mod debug_as_type;
+pub(crate) use debug_as_type::DebugAsType;
+
 pub mod metrics;
 #[cfg(feature = "legacy")]
 pub use metrics::types::BackgroundMetrics;
